Made a component responsible for one action
