+++ src/components/ExpenseList.tsx
@@ -1,5 +1,3 @@
-import { ChangeEvent, useState } from "react";
-
 interface Expense {
   id: string;
   description: string;
@@ -8,25 +6,13 @@ interface Expense {
 }
 
 interface Props {
-  expenses: Expense[];
   setExpenses: (expenses: Expense[]) => void;
   onDelete: (id: string) => void;
+  filteredExpense: Expense[];
 }
 
-function Category({ expenses, onDelete }: Props) {
-  const [selectedCategory, setSelectedCategory] = useState("all");
-
+function ExpenseList({ onDelete, filteredExpense }: Props) {
   let total = 0;
-  const handleOnChange = (e: ChangeEvent<HTMLSelectElement>) => {
-    setSelectedCategory(e.target.value);
-  };
-
-  const filteredExpense = expenses.filter((expense: Expense) =>
-    selectedCategory === "all"
-      ? expense.category
-      : expense.category === selectedCategory
-  );
-
   filteredExpense.map((obj) => (total += Number(obj.amount)));
 
   const expenseList = filteredExpense.map((obj: Expense) => (
@@ -47,17 +33,6 @@ function Category({ expenses, onDelete }: Props) {
 
   return (
     <div>
-      <select
-        name="expenses"
-        className="form-select form-select-lg mb-3"
-        value={selectedCategory}
-        onChange={handleOnChange}
-      >
-        <option value="all">All categories</option>
-        <option value="Entertainment">Entertainment</option>
-        <option value="Utilities">Utilities</option>
-        <option value="Groceries">Groceries</option>
-      </select>
       <table className="table table-bordered">
         <thead>
           <tr>
@@ -79,4 +54,4 @@ function Category({ expenses, onDelete }: Props) {
   );
 }
 
-export default Category;
+export default ExpenseList;
